Pull SystemCard out into its own component
